Getting file chunks

+++ utils/piece_manager.go
@@ -13,16 +13,16 @@ const (
 )
 
 type PieceManager struct {
-	TotalPieces int
-	PieceLength int
-	TotalLength int
+	TotalPieces int64
+	PieceLength int64
+	TotalLength int64
 
 	MyBitfield BitField
 	Statuses   []PieceStatus
 	Mu         sync.Mutex
 }
 
-func NewPieceManager(totalLength, pieceLength int) *PieceManager {
+func NewPieceManager(totalLength, pieceLength int64) *PieceManager {
 	numPieces := (totalLength + pieceLength - 1) / pieceLength
 	return &PieceManager{
 		TotalPieces: numPieces,
@@ -37,7 +37,7 @@ func (pm *PieceManager) PickPiece(peerBf BitField) (int, bool) {
 	pm.Mu.Lock()
 	defer pm.Mu.Unlock()
 
-	for i := 0; i < pm.TotalPieces; i++ {
+	for i := 0; i < int(pm.TotalPieces); i++ {
 		if pm.Statuses[i] == PieceNotStarted && peerBf.HasPiece(i) {
 			pm.Statuses[i] = PieceDownloading
 			return i, true
